pyproject.toml
[project]
name = "mcpfire"
version = "0.0.1"
description = "Turn any REST API into an MCP server in minutes - no code, no deployments. MCPFire is a Meta-MCP server that manages all your tools."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Vitalii Stepanenko", email = "mail@vitaliy.in" },
]
maintainers = [
    { name = "Vitalii Stepanenko", email = "mail@vitaliy.in" },
]
license = { file = "LICENSE" }
dependencies = [
    "ai-microcore~=4.4",
    "cyclopts~=4.2",
    "hypercorn~=0.18",
    "config_reading[pyyaml]~=0.1",
    "fastapi>=0.122,<1",
    "pydantic>=2.12.4,<3",
]

[dependency-groups]
dev = [
    "pytest~=9.0",
    "pytest-asyncio>=1.3.0",
    "flake8-pyproject"
]

[project.urls]
"Source Code" = "https://github.com/Nayjest/MCP-Fire"


[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["mcpfire"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.flake8]
max-line-length = 100

[project.scripts]
mcpfire = "mcpfire.cli:cli"